Add Observable Concept
This is optional. Any class can use the Observable<T> type instead of a primitive type (e.g. string, float, int) and it comes with OnChanged event broadcast upon changes and a nice inspector. Its particularly use for the Model in UMVCS because non-Model scopes want to know when the model changes.